Move cube generation under utilities
