+++ cpp/CMakeLists.txt
@@ -148,7 +148,6 @@ endif()
 add_library(chemistry)
 add_subdirectory(src/qdk/chemistry/data)
 add_subdirectory(src/qdk/chemistry/algorithms)
-add_subdirectory(src/qdk/chemistry/cube)
 add_subdirectory(src/qdk/chemistry/utils)
 
 target_include_directories(chemistry PUBLIC
