--- cpp/src/qdk/chemistry/cube/CMakeLists.txt
@@ -1,6 +1,0 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License.
-
-target_sources(chemistry PRIVATE
-    cube_generator.cpp
-)
