+++ cpp/src/qdk/chemistry/utils/CMakeLists.txt
@@ -1,4 +1,5 @@
 target_sources(chemistry PRIVATE
+    cube_generator.cpp
     hash_context.cpp
     valence_space.cpp
     orbital_rotation.cpp
